App-3-Project-Math-Painting/requirements.txt
numpy==2.0.0
Pillow==10.3
